+++ cashu-lib-crypto/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.14.0] - 2026-01-28
+
+### Changed
+
+- Version sync with parent cashu-lib 0.14.0
+
+---
+
 ## [0.12.0] - 2026-01-21
 
 ### Added